Refactor to use strategy pattern
When applied, this commit will create a class for each condition within the if statement on service class.
It is responsible to lower coupling by injecting the strategy class created by strategy factory on calculate shipping method on shipping service class.
The creation of a class for each if clause helps to encapsulate business rules and to respect Single Responsability Principle.